Add a static image file
We extend the header template to add our own logo to the site. Static
files like images are stored on the public folder that we registered
initially.

We could link to the image directly in our template using
/img/dsm-logo.jpg, but is safer to use the url_for_static helper as it
takes into account things like whether CKAN is mounted on a non-root
URL (like example.com/data)